style: fix ESLint errors and warnings
