+++ .github/workflows/lint.yml
@@ -1,0 +1,51 @@
+name: Lint
+
+on:
+  push:
+    branches:
+      - vue3
+  pull_request:
+    branches:
+      - vue3
+  workflow_dispatch:
+
+jobs:
+  eslint:
+    name: ESLint
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2.3.4
+      - name: Cache NPM packages
+        uses: actions/cache@v2.1.6
+        id: npm-cache
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+      - name: Install dependencies
+        run: npm ci --no-audit
+      - name: Run ESLint
+        run: npm run lint:js
+
+  stylelint:
+    name: Stylelint
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2.3.4
+      - name: Cache NPM packages
+        uses: actions/cache@v2.1.6
+        id: npm-cache
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+      - name: Install dependencies
+        run: npm ci --no-audit
+      - name: Run Stylelint
+        run: npm run lint:style
