--- components.d.ts
@@ -1,24 +1,0 @@
-// generated by vite-plugin-components
-// read more https://github.com/vuejs/vue-next/pull/3399
-
-declare module 'vue' {
-  export interface GlobalComponents {
-    MediaInfo: typeof import('./src/components/media-info.vue')['default']
-    PlayButton: typeof import('./src/components/buttons/play-button.vue')['default']
-    BlurhashCanvas: typeof import('./src/components/images/blurhash-canvas.vue')['default']
-    BlurhashImage: typeof import('./src/components/images/blurhash-image.vue')['default']
-    DarkModeToggle: typeof import('./src/components/system/dark-mode-toggle.vue')['default']
-    LocaleSwitcher: typeof import('./src/components/system/locale-switcher.vue')['default']
-    Header: typeof import('./src/components/layout/home/header/header.vue')['default']
-    AlbumTitle: typeof import('./src/components/layout/home/header/items/album-title.vue')['default']
-    EpisodeTitle: typeof import('./src/components/layout/home/header/items/episode-title.vue')['default']
-    GenericTitle: typeof import('./src/components/layout/home/header/items/generic-title.vue')['default']
-    HelloWorld: typeof import('./src/components/HelloWorld.vue')['default']
-    HomeHeader: typeof import('./src/components/layout/home/headerSection/HomeHeader.vue')['default']
-    HomeHeaderAlbumTitle: typeof import('./src/components/layout/home/headerSection/items/HomeHeaderAlbumTitle.vue')['default']
-    HomeHeaderEpisodeTitle: typeof import('./src/components/layout/home/headerSection/items/HomeHeaderEpisodeTitle.vue')['default']
-    HomeHeaderGenericTitle: typeof import('./src/components/layout/home/headerSection/items/HomeHeaderGenericTitle.vue')['default']
-  }
-}
-
-export { }
