+++ src/components/layout/home/header/items/generic-title.vue
@@ -30,42 +30,25 @@
   </div>
 </template>
 
-<script setup lang="ts">
-import { defineProps } from 'vue';
+<script lang="ts">
 import type { BaseItemDto } from '@jellyfin/client-axios';
+import { defineComponent } from 'vue';
 import { useDisplay } from 'vuetify/lib/composables/display';
+
 import { useBaseItem } from '~/composables/items';
 
-const props = defineProps({
-  item: {
-    type: Object as () => BaseItemDto,
-    required: true
-  }
-});
-
-const display = useDisplay();
-
-const { getLogo } = useBaseItem();
-
-//import imageHelper from '~/mixins/imageHelper';
-//import itemHelper from '~/mixins/itemHelper';
-/*
 export default defineComponent({
-  //mixins: [imageHelper, itemHelper],
   props: {
     item: {
       type: Object as () => BaseItemDto,
       required: true
-    },
-    logo: {
-      type: String,
-      default: ''
     }
+  },
+  setup() {
+    const display = useDisplay();
+    const { getLogo } = useBaseItem();
+
+    return { display, getLogo };
   }
-  /*computed: {
-    imageTag(): string | undefined {
-      return this.getImageTag(this.item, ImageType.Logo);
-    }
-  }*/
-//}); */
+});
 </script>
