+++ src/vite-env.d.ts
@@ -1,1 +1,2 @@
 /// <reference types="vite/client" />
+// eslint-disable-next-line unicorn/prevent-abbreviations -- Keep the standard file name
